refactor: move room related api call to its own file
